+++ jovo-platforms/jovo-platform-googleassistantconv/src/GoogleAssistant.ts
@@ -56,6 +56,7 @@ export class GoogleAssistant extends Platform<
     app.middleware('platform.nlu')!.use(this.nlu.bind(this));
     app.middleware('platform.output')!.use(this.output.bind(this));
     app.middleware('response')!.use(this.response.bind(this));
+    app.middleware('after.response')!.use(this.afterResponse.bind(this));
 
     Jovo.prototype.$googleAction = undefined;
 
@@ -149,6 +150,12 @@ export class GoogleAssistant extends Platform<
     handleRequest.jovo!.$response = handleRequest.jovo!.$rawResponseJson
       ? this.responseBuilder.create(handleRequest.jovo!.$rawResponseJson)
       : handleRequest.jovo!.$response;
+  }
+
+  async afterResponse(handleRequest: HandleRequest) {
+    if (!isValidGoogleAssistantRequest(handleRequest)) {
+      return Promise.resolve();
+    }
     await handleRequest.host.setResponse(handleRequest.jovo!.$response);
   }