+++ jovo-platforms/jovo-platform-twilioautopilot/src/Autopilot.ts
@@ -61,6 +61,7 @@ export class Autopilot extends Platform<AutopilotRequest, AutopilotResponse> {
     app.middleware('tts')!.use(this.tts.bind(this));
     app.middleware('platform.output')!.use(this.output.bind(this));
     app.middleware('response')!.use(this.response.bind(this));
+    app.middleware('after.response')!.use(this.afterResponse.bind(this));
 
     this.use(new AutopilotCore(), new AutopilotNLU(), new AudioPlayerPlugin(), new Cards());
 
@@ -129,6 +130,12 @@ export class Autopilot extends Platform<AutopilotRequest, AutopilotResponse> {
       return Promise.resolve();
     }
     await this.middleware('$response')!.run(handleRequest.jovo);
+  }
+
+  async afterResponse(handleRequest: HandleRequest) {
+    if (handleRequest.jovo?.constructor.name !== 'AutopilotBot') {
+      return Promise.resolve();
+    }
 
     await handleRequest.host.setResponse(handleRequest.jovo.$response);
   }