+++ jovo-integrations/jovo-analytics-dashbot/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Current version might be higher than the latest changes displayed below because of updates of dependencies.
 
+### 2021-02-22 [3.5]
+- [#898](https://github.com/jovotech/jovo-framework/pull/898) Add error handling to analytics integrations ([@techpeace](https://github.com/techpeace)
+
 
 ### 2020-08-21  [3.0.26]